# docs/swagger.yaml
basePath: /
definitions:
...
  ent.FLink:
    properties:
      avatar_url:
        description: 头像
        type: string
      cover_url:
        description: 网站封面
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 简介
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.FLinkEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FLinkQuery when eager-loading is set.
      email:
        description: 邮箱
        type: string
      enable_friend_circle:
        description: 是否开启朋友圈
        type: boolean
      friend_circle_rule_id:
        description: 朋友圈解析规则
        type: integer
      group_id:
        description: 友链组
        type: integer
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 名称
        type: string
      snapshot_url:
        description: 网站截图
        type: string
      status:
        description: 状态
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      url:
        description: 链接
        type: string
    type: object
  ent.FLinkEdges:
    properties:
      group:
        allOf:
        - $ref: '#/definitions/ent.FLinkGroup'
        description: Group holds the value of the group edge.
    type: object
  ent.FLinkGroup:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.FLinkGroupEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FLinkGroupQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 名称
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.FLinkGroupEdges:
    properties:
      links:
        description: Links holds the value of the links edge.
        items:
          $ref: '#/definitions/ent.FLink'
        type: array
    type: object
...
  model.EssayCreateReq:
    properties:
      content:
        type: string
      draft:
        type: boolean
      images:
        items:
          type: string
        type: array
    type: object
  model.EssayResp:
    properties:
      content:
        type: string
      create_at:
        type: string
      draft:
        type: boolean
      id:
        type: integer
      images:
        items:
          type: string
        type: array
    type: object
  model.EssayUpdateReq:
    properties:
      content:
        type: string
      draft:
        type: boolean
      id:
        type: integer
      images:
        items:
          type: string
        type: array
    type: object
  model.FlinkCreateReq:
    properties:
      avatar_url:
        description: logo
        type: string
      cover_url:
        type: string
      description:
        description: 简介
        type: string
      email:
        description: 邮箱
        type: string
      enable_friend_circle:
        description: 是否开启朋友圈
        type: boolean
      friend_circle_rule_id:
        description: 朋友圈解析规则
        type: integer
      name:
        description: 名称
        type: string
      snapshot_url:
        description: 快照
        type: string
      url:
        description: 链接
        type: string
    type: object
  model.FlinkResp:
    properties:
      avatar_url:
        description: logo
        type: string
      cover_url:
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 简介
        type: string
      email:
        description: 邮箱
        type: string
      enable_friend_circle:
        description: 是否开启朋友圈
        type: boolean
      friend_circle_rule_id:
        description: 朋友圈解析规则
        type: integer
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 名称
        type: string
      snapshot_url:
        description: 快照
        type: string
      status:
        description: 状态
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      url:
        description: 链接
        type: string
    type: object
  model.FlinkUpdateReq:
    properties:
      avatar_url:
        description: logo
        type: string
      cover_url:
        type: string
      description:
        description: 简介
        type: string
      email:
        description: 邮箱
        type: string
      enable_friend_circle:
        description: 是否开启朋友圈
        type: boolean
      friend_circle_rule_id:
        description: 朋友圈解析规则
        type: integer
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 名称
        type: string
      snapshot_url:
        description: 快照
        type: string
      url:
        description: 链接
        type: string
    type: object
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.PageResult-model_EssayResp:
    properties:
      records:
        items:
          $ref: '#/definitions/model.EssayResp'
        type: array
      total:
        type: integer
    type: object
  model.PageResult-model_FlinkResp:
    properties:
      records:
        items:
          $ref: '#/definitions/model.FlinkResp'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/essay/create:
    post:
      consumes:
      - application/json
      description: 创建一个新说说
      parameters:
      - description: 说创建请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.EssayCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.EssayCreateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建说说
      tags:
      - essay
  /api/v1/essay/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定说说
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.EssayResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除说说
      tags:
      - essay
  /api/v1/essay/page:
    get:
      consumes:
      - application/json
      description: 获取说说分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-model_EssayResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取说说分页列表
      tags:
      - essay
  /api/v1/essay/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定说说的信息
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: string
      - description: 说说更新请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.EssayUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.EssayUpdateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新说说
      tags:
      - essay
...
  /api/v1/flink/{id}:
    delete:
      consumes:
      - application/json
      description: 删除Flink
      parameters:
      - description: Flink ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除Flink
      tags:
      - flink
  /api/v1/flink/create:
    post:
      consumes:
      - application/json
      description: 创建Flink
      parameters:
      - description: Flink创建请求体
        in: body
        name: flink_create_req
        required: true
        schema:
          $ref: '#/definitions/model.FlinkCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.FLink'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建Flink
      tags:
      - flink
  /api/v1/flink/list:
    get:
      consumes:
      - application/json
      description: 获取所有Flink
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.FlinkResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取所有Flink
      tags:
      - flink
  /api/v1/flink/page:
    get:
      consumes:
      - application/json
      description: 获取Flink分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-model_FlinkResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取Flink分页列表
      tags:
      - flink
  /api/v1/flink/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询Flink
      parameters:
      - description: Flink ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.FLink'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询Flink
      tags:
      - flink
  /api/v1/flink/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新Flink
      parameters:
      - description: Flink ID
        in: path
        name: id
        required: true
        type: integer
      - description: Flink更新请求体
        in: body
        name: flink_update_req
        required: true
        schema:
          $ref: '#/definitions/model.FlinkUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.FLink'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新Flink
      tags:
      - flink
...
swagger: "2.0"